requirements.txt
feedparser
newspaper3k
requests
streamlit
pandas
python-dotenv
lxml_html_clean
